fix(pr): retain observed stack bases and extract GitHub transport
